feat: Migrate Cursor extensions plugin to mixin architecture
**Cursor Extensions Migration:**
- 57% code reduction: 361 → 155 lines
- Migrated to ExtensionsPlugin<CursorExtensionsCore>
- Uses "cursor" command with --list-extensions --show-versions
- Full restore functionality with alternative file detection

**Migration Progress:**
- 5 plugins now migrated with consistent results:
  - VSCode settings: 53% reduction
  - VSCode extensions: 58% reduction
  - Cursor settings: 46% reduction
  - Cursor extensions: 57% reduction

**Average Code Reduction: 54%** - exceeding targets\!

**Proven Architecture Benefits:**
- Extreme code reuse through mixin composition
- Zero duplication of common patterns
- Type-safe configuration and validation
- Consistent restore logic with dry-run support
- Enhanced error handling and user feedback

Ready for package managers (Homebrew, NPM) which should show
even greater reductions due to their complexity.
